Fix wrongly translated captions
In the german version of the template captions are wrongly translated. For example "Contents" instead of "Inhaltsverzeichnis" and so on.

Babel gave the following warning:
Package babel Warning: Last declared language option is `ngerman',
(babel)                but the last processed one was `american'.
(babel)                The main language cannot be set as both a global
(babel)                and a package option. Use `main=ngerman' as
(babel)                option. Reported on input line 460.

Using `main=ngerman' fixes the problem.